Replace Rewrite.Project by Rewrite
